+++ src/pagination/index.ts
@@ -1,4 +1,2 @@
-export * from './BasePaginator';
-export * from './ChannelPaginator';
+export * from './paginators';
 export * from './FilterBuilder';
-export * from './ReminderPaginator';
